Fix clang compiler failures

+++ src/umpire/event/json_file_store.cpp
@@ -38,9 +38,9 @@ void json_file_store::insert(const allocate& e)
 {
   fprintf(m_fstream,
           "{\"category\":\"operation\",\"name\":\"allocate\",\"numeric_args\":{\"size\":%ld},\"string_args\":{"
-          "\"allocator_ref\":\"%p\",\"pointer\":\"%p\"},\"tags\":{\"replay\":\"true\"},\"timestamp\":%ld}\n",
+          "\"allocator_ref\":\"%p\",\"pointer\":\"%p\"},\"tags\":{\"replay\":\"true\"},\"timestamp\":%lld}\n",
           e.size, e.ref, e.ptr,
-          std::chrono::time_point_cast<std::chrono::nanoseconds>(e.timestamp).time_since_epoch().count());
+          static_cast<long long>(std::chrono::time_point_cast<std::chrono::nanoseconds>(e.timestamp).time_since_epoch().count()));
 }
 
 void json_file_store::insert(const named_allocate& e)
@@ -48,17 +48,17 @@ void json_file_store::insert(const named_allocate& e)
   fprintf(m_fstream,
           "{\"category\":\"operation\",\"name\":\"named_allocate\",\"numeric_args\":{\"size\":%ld},\"string_args\":{"
           "\"allocator_ref\":\"%p\",\"pointer\":\"%p\",\"allocation_name\":\"%s\"},\"tags\":{\"replay\":\"true\"},"
-          "\"timestamp\":%ld}\n",
+          "\"timestamp\":%lld}\n",
           e.size, e.ref, e.ptr, e.name.c_str(),
-          std::chrono::time_point_cast<std::chrono::nanoseconds>(e.timestamp).time_since_epoch().count());
+          static_cast<long long>(std::chrono::time_point_cast<std::chrono::nanoseconds>(e.timestamp).time_since_epoch().count()));
 }
 
 void json_file_store::insert(const deallocate& e)
 {
   fprintf(m_fstream,
           "{\"category\":\"operation\",\"name\":\"deallocate\",\"string_args\":{\"allocator_ref\":\"%p\",\"pointer\":"
-          "\"%p\"},\"tags\":{\"replay\":\"true\"},\"timestamp\":%ld}\n",
-          e.ref, e.ptr, std::chrono::time_point_cast<std::chrono::nanoseconds>(e.timestamp).time_since_epoch().count());
+          "\"%p\"},\"tags\":{\"replay\":\"true\"},\"timestamp\":%lld}\n",
+          e.ref, e.ptr, static_cast<long long>(std::chrono::time_point_cast<std::chrono::nanoseconds>(e.timestamp).time_since_epoch().count()));
 }
 
 std::vector<event> json_file_store::get_events()
